Auto-open and auto-listen
This is a largely untested feature that gives the user an option to associate a Profile stored within the Profiles sub-folder at GAVPI's location with an executable file and, when that executable starts, automatically open the associated Profile (with the option to begin Listening for voice commands).  This feature has only been test on Windows 8.1 64bit.

Modified frmProfile.Designer.cs to offer UI elements allowing a user to select an executable file to associate with the current Profile.

Updated VI_Profile.cs to add an additional XML node (AssociatedProcess) within which an executable file associated with the Profile is persisted.  Added getter/setter for the associated process.

Updated Settings.Designer.cs to allow the option of automatically loading Profiles when an associated executable starts, along with the option to automatically begin listening for voice commands.  These options are persisted through the applications own Properties.Settings.Default... settings.

Implemented a ManagementEventWatcher and WlqEventQuery within GAVPI.cs, via EnableAutoOpenProfile() and DisableAutoOpenProfile(), to start and stop watching for process startup, matching processes with an executable associated with a Profile.

Fixed an oversight with GAVPI.OpenSettings().